+++ book-room/quarto/src/main/java/br/com/book/room/quarto/domain/core/QuartoCamaId.java
@@ -2,5 +2,5 @@ package br.com.book.room.quarto.domain.core;
 
 import java.io.Serializable;
 
-public record QuartoCamaId(Integer id, Integer tipoCama) implements Serializable {
+public record QuartoCamaId(Long id, Long tipoCama) implements Serializable {
 }
